Final submit for basic To do list

+++ js/main.js
@@ -1,4 +1,7 @@
 var toDoTask = new ToDoTasks();
+var validate = new Validation();
+
+getLocalStorage();
 
 function getElement(input) {
   return document.getElementById(input);
@@ -8,11 +11,32 @@ getElement("addItem").addEventListener("click", function () {
   var taskName = getElement("newTask").value;
   var id = Math.floor(Math.random() * 100);
   var status = "todo";
+  var isEmpty = true;
+  var isDuplicated;
+
+  isEmpty = validate.checkEmpty(
+    taskName,
+    "notiInput",
+    "(*) Task can't leave blank"
+  );
+  console.log(isEmpty);
+
+  isDuplicated = validate.checkDuplicateTask(
+    toDoTask.arr,
+    taskName,
+    "duplicateInput",
+    "(*) This task has already been existed"
+  );
+  console.log(isDuplicated);
+
+  if (!isEmpty || !isDuplicated) return;
 
   var task = new Task(id, taskName, status);
   toDoTask.addTask(task);
 
   createToDoList(toDoTask.arr);
+  setLocalStorage();
+  getElement("newTask").value = "";
 });
 
 function createToDoList(arr) {
@@ -28,9 +52,9 @@ function createToDoList(arr) {
   `;
     } else {
       completedContent += `
-      <li><p>${item.taskName}</p>
+      <li><p style="color:#25b99a">${item.taskName}</p>
           <button id="deleteButton" onclick="deleteButton('${item.id}')"><i class="fa fa-trash"></i></button>
-          <button id="checkButton" onclick="updateButton('${item.id}')"><i class="fa fa-check"></i></button>
+          <button id="checkButton" onclick="updateButton('${item.id}')" style="color:#25b99a"><i class="fa fa-check"></i></button>
       </li>
   `;
     }
@@ -42,8 +66,13 @@ function createToDoList(arr) {
 }
 
 function deleteButton(id) {
-  toDoTask.deleteTask(id);
-  createToDoList(toDoTask.arr);
+  var answer = confirm("Do you wanna delete this task?");
+  if (answer == true) {
+    toDoTask.deleteTask(id);
+    createToDoList(toDoTask.arr);
+    alert("Delete Successfully");
+    setLocalStorage();
+  }
 }
 
 function updateButton(id) {
@@ -51,5 +80,17 @@ function updateButton(id) {
 
   taskDetail.status = "todo" === taskDetail.status ? "completed" : "todo";
   toDoTask.updateStatus(taskDetail);
+  alert("Update status successfully");
   createToDoList(toDoTask.arr);
+}
+
+function getLocalStorage() {
+  if (localStorage.getItem("ToDoList")) {
+    toDoTask.arr = JSON.parse(localStorage.getItem("ToDoList"));
+    createToDoList(toDoTask.arr);
+  }
+}
+
+function setLocalStorage() {
+  localStorage.setItem("ToDoList", JSON.stringify(toDoTask.arr));
 }
